# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Loan CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List clients (filtered by role automatically)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [ativo, em_cobranca, quitado]
        - name: vendedor_id
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a client and auto-generate installments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Client created with installments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get client with full details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientDetail"
        "404":
          description: Not found
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdate"
      responses:
        "200":
          description: Updated client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}/risk-level:
    patch:
      operationId: updateClientRiskLevel
      tags: [clients]
      summary: Manually set or clear client risk level
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RiskLevelInput"
      responses:
        "200":
          description: Updated client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"

  # INSTALLMENTS
  /installments:
    get:
      operationId: listInstallments
      tags: [installments]
      summary: List installments with filters
      parameters:
        - name: client_id
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pendente, pago, atrasado]
        - name: vendedor_id
          in: query
          schema:
            type: integer
        - name: overdue_only
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of installments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Installment"
  /installments/{id}:
    get:
      operationId: getInstallment
      tags: [installments]
      summary: Get installment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Installment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Installment"
        "404":
          description: Not found
    patch:
      operationId: updateInstallment
      tags: [installments]
      summary: Update installment (mark as paid, update value, etc.)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstallmentUpdate"
      responses:
        "200":
          description: Updated installment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Installment"
  /installments/{id}/pay:
    post:
      operationId: payInstallment
      tags: [installments]
      summary: Mark installment as paid and create event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "200":
          description: Payment registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Installment"
  /clients/{id}/renovacao:
    post:
      operationId: renovacaoClient
      tags: [installments]
      summary: Generate new installments for renewal/agreement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenovacaoInput"
      responses:
        "200":
          description: Renewal registered with new installments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"

  # CLIENT EVENTS
...
components:
  schemas:
...
    Client:
      type: object
      required: [id, nome, telefone, vendedor_id, valor_contrato, parcelas, dia_vencimento, status]
      properties:
        id:
          type: integer
        nome:
          type: string
        telefone:
          type: string
        vendedor_id:
          type: integer
        vendedor_nome:
          type: ["string", "null"]
        valor_contrato:
          type: number
        parcelas:
          type: integer
        dia_vencimento:
          type: integer
        status:
          type: string
          enum: [ativo, em_cobranca, quitado]
        risk_level:
          type: ["string", "null"]
          enum: [atencao, risco, critico, null]
        createdAt:
          type: string
    ClientDetail:
      type: object
      required: [id, nome, telefone, vendedor_id, valor_contrato, parcelas, dia_vencimento, status, installments, events]
      properties:
        id:
          type: integer
        nome:
          type: string
        telefone:
          type: string
        vendedor_id:
          type: integer
        vendedor_nome:
          type: ["string", "null"]
        valor_contrato:
          type: number
        parcelas:
          type: integer
        dia_vencimento:
          type: integer
        status:
          type: string
          enum: [ativo, em_cobranca, quitado]
        risk_level:
          type: ["string", "null"]
          enum: [atencao, risco, critico, null]
        createdAt:
          type: string
        installments:
          type: array
          items:
            $ref: "#/components/schemas/Installment"
        events:
          type: array
          items:
            $ref: "#/components/schemas/ClientEvent"
    RiskLevelInput:
      type: object
      required: [risk_level]
      properties:
        risk_level:
          type: ["string", "null"]
          enum: [atencao, risco, critico, null]
    ClientInput:
      type: object
      required: [nome, telefone, valor_contrato, parcelas, dia_vencimento]
      properties:
        nome:
          type: string
        telefone:
          type: string
        valor_contrato:
          type: number
        parcelas:
          type: integer
        dia_vencimento:
          type: integer
        vendedor_id:
          type: ["integer", "null"]
    ClientUpdate:
      type: object
      properties:
        nome:
          type: string
        telefone:
          type: string
        status:
          type: string
          enum: [ativo, em_cobranca, quitado]
        valor_contrato:
          type: number

    # INSTALLMENTS
    Installment:
      type: object
      required: [id, client_id, numero_parcela, valor, vencimento, status]
      properties:
        id:
          type: integer
        client_id:
          type: integer
        client_nome:
          type: ["string", "null"]
        vendedor_id:
          type: ["integer", "null"]
        numero_parcela:
          type: integer
        valor:
          type: number
        vencimento:
          type: string
        status:
          type: string
          enum: [pendente, pago, atrasado]
        pago_em:
          type: ["string", "null"]
    InstallmentUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pendente, pago, atrasado]
        valor:
          type: number
        vencimento:
          type: string
    PaymentInput:
      type: object
      required: [valor_pago]
      properties:
        valor_pago:
          type: number
        observacao:
          type: string
    RenovacaoInput:
      type: object
      required: [novas_parcelas, novo_valor, novo_dia_vencimento]
      properties:
        novas_parcelas:
          type: integer
        novo_valor:
          type: number
        novo_dia_vencimento:
          type: integer
        observacao:
          type: string

    # EVENTS
    ClientEvent:
      type: object
      required: [id, client_id, tipo, user_id, data]
      properties:
        id:
          type: integer
        client_id:
          type: integer
        client_nome:
          type: ["string", "null"]
        tipo:
          type: string
          enum: [venda_fechada, parcela_paga, atraso, renovacao, follow_up, anotacao]
        user_id:
          type: integer
        user_nome:
          type: ["string", "null"]
        observacao:
          type: ["string", "null"]
        data:
          type: string
...