# lib/api-zod/src/generated/types/client.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Loan CRM API
 * OpenAPI spec version: 0.1.0
 */
import type { ClientRiskLevel } from './clientRiskLevel';
import type { ClientStatus } from './clientStatus';

export interface Client {
  id: number;
  nome: string;
  telefone: string;
  vendedor_id: number;
  /** @nullable */
  vendedor_nome?: string | null;
  valor_contrato: number;
  parcelas: number;
  dia_vencimento: number;
  status: ClientStatus;
  /** @nullable */
  risk_level?: ClientRiskLevel;
  createdAt?: string;
}