# lib/api-zod/src/generated/types/clientDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Loan CRM API
 * OpenAPI spec version: 0.1.0
 */
import type { ClientDetailRiskLevel } from './clientDetailRiskLevel';
import type { ClientDetailStatus } from './clientDetailStatus';
import type { ClientEvent } from './clientEvent';
import type { Installment } from './installment';

export interface ClientDetail {
  id: number;
  nome: string;
  telefone: string;
  vendedor_id: number;
  /** @nullable */
  vendedor_nome?: string | null;
  valor_contrato: number;
  parcelas: number;
  dia_vencimento: number;
  status: ClientDetailStatus;
  /** @nullable */
  risk_level?: ClientDetailRiskLevel;
  createdAt?: string;
  installments: Installment[];
  events: ClientEvent[];
}