# lib/api-zod/src/generated/types/riskLevelInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Loan CRM API
 * OpenAPI spec version: 0.1.0
 */
import type { RiskLevelInputRiskLevel } from './riskLevelInputRiskLevel';

export interface RiskLevelInput {
  /** @nullable */
  risk_level: RiskLevelInputRiskLevel;
}